upgrade example to CPAL 0.1.2
The signed/unsigned bug in CPAL has been fixed!

+++ examples/cpal.rs
@@ -33,15 +33,13 @@ fn main() {
 
         if samples_left == 0 { return false; }
 
-        let mut buffer: cpal::Buffer<u16> =
+        let mut buffer: cpal::Buffer<i16> =
             voice.append_data(spec.channels,
                               cpal::SamplesRate(spec.sample_rate),
                               samples_left);
         // Fill the cpal buffer with data from the wav file.
         for (dest, src) in buffer.iter_mut().zip(&mut samples) {
-            // TODO: There is a bug in cpal that handles signed samples in the
-            // wrong manner, so we cast it to `u16` for now.
-            *dest = src.unwrap() as u16;
+            *dest = src.unwrap();
         }
 
         // Probably not done, loop again.
